refactor(gridsat): consolidate GSDatasetInfo import and remove netcdf_metadata.py
