+++ src/goesdl/gridsat/__init__.py
@@ -8,9 +8,9 @@ Classes:
 
 from .locator_b1 import GridSatProductLocatorB1
 from .locator_gc import GridSatProductLocatorGC
+from .netcdf_dataset import GSDatasetInfo
 from .netcdf_geodetic import GSLatLonGrid
 from .netcdf_image import GSImage
-from .netcdf_metadata import GSDatasetInfo
 from .netcdf_time import GSCoverageTime, GSTimeGrid
 from .utility import read_gridsat_dataset
