--- src/goesdl/gridsat/netcdf_metadata.py
@@ -1,102 +1,0 @@
-import math
-from typing import Any
-
-from netCDF4 import Dataset
-
-from ..netcdf import HasStrHelp
-from .constants import NA
-from .databook_gc import (
-    PIXELS_PER_DEGREE,
-    abstract_gridsat_gc,
-    channel_correspondence,
-    channel_description_gc,
-    dataset_name_gridsat_gc,
-    geospatial_resolution_deg,
-    geospatial_resolution_km,
-    measurement_range_lower_bound,
-    measurement_range_upper_bound,
-    measurement_units,
-    platform_origin_gridsat_gc,
-    square_igfov_at_nadir,
-    wavelength_range_lower_bound,
-    wavelength_range_upper_bound,
-)
-from .netcdf_dataset import GSDatasetMetadata
-from .validation_gc import (  # pylint: disable=no-name-in-module
-    validate_channel,
-)
-
-NAN_TUPLE = math.nan, math.nan
-
-
-class DatabookMetadata(HasStrHelp):
-
-    dataset: str = NA
-    abstract: str = NA
-    description: str = NA
-    channel_id: str = NA
-    channel: str = NA
-    geospatial_resolution_deg: tuple[float, float] = NAN_TUPLE
-    geospatial_resolution_km: tuple[float, float] = NAN_TUPLE
-    measurement_bounds: tuple[float, float] = NAN_TUPLE
-    measurement_units: str = NA
-    pixels_per_degree: int = PIXELS_PER_DEGREE
-    square_fov_at_nadir: float = math.nan
-    wavelength: float = math.nan
-    wavelength_bounds: tuple[float, float] = NAN_TUPLE
-
-    def __init__(self, channel: str, platform: str) -> None:
-        # Validate platform parameter
-        if platform not in platform_origin_gridsat_gc:
-            allowed_platforms = ", ".join(platform_origin_gridsat_gc.keys())
-            raise ValueError(
-                f"Invalid 'platform': '{platform}'; "
-                f"allowed platforms are: {allowed_platforms}"
-            )
-
-        origin = platform_origin_gridsat_gc[platform]
-        channel_orig = channel_correspondence[origin][channel]
-
-        self.dataset = dataset_name_gridsat_gc
-        self.abstract = abstract_gridsat_gc
-
-        if channel_orig == 0:
-            self.channel += (
-                f"GridSat '{channel}' is not supported by {platform} origin"
-            )
-            return
-
-        self.description = channel_description_gc[channel]
-
-        self.channel_id = channel
-        self.channel = f"Channel {channel_orig}"
-
-        self.geospatial_resolution_deg = geospatial_resolution_deg
-        self.geospatial_resolution_km = geospatial_resolution_km
-
-        measurement_lo = measurement_range_lower_bound[origin][channel_orig]
-        measurement_up = measurement_range_upper_bound[origin][channel_orig]
-        self.measurement_bounds = measurement_lo, measurement_up
-
-        self.measurement_units = measurement_units[channel_orig]
-
-        wavelength_lo = wavelength_range_lower_bound[origin][channel_orig]
-        wavelength_up = wavelength_range_upper_bound[origin][channel_orig]
-        self.wavelength_bounds = wavelength_lo, wavelength_up
-
-        self.wavelength = 0.5 * (wavelength_lo + wavelength_up)
-
-        self.square_fov_at_nadir = square_igfov_at_nadir[origin][channel_orig]
-
-
-class GSDatasetInfo(DatabookMetadata, GSDatasetMetadata):
-
-    def __init__(self, record: Dataset, channel: str) -> None:
-        validate_channel(channel, record)
-
-        GSDatasetMetadata.__init__(self, record, channel=channel)
-
-    def __post_init__(self, record: Dataset, **kwargs: Any) -> None:
-        channel: str = kwargs["channel"]
-
-        DatabookMetadata.__init__(self, channel, self.platform)
