+++ src/goesdl/gridsat/utility.py
@@ -3,8 +3,8 @@ from netCDF4 import Dataset
 from ..geodesy import RectangularRegion
 from ..gridsat.netcdf_geodetic import GSLatLonGrid
 from ..gridsat.netcdf_image import GSImage
-from ..gridsat.netcdf_metadata import GSDatasetInfo
 from ..gridsat.netcdf_time import GSCoverageTime
+from .netcdf_dataset import GSDatasetInfo
 
 
 def read_gridsat_dataset(
